Refactor: Chatting Retrofit Refactoring
- Chatting.Retrofit에 있던 매칭수동마감 api -> ChattingRoom.Retrofit으로 이동 후, Chatting.Retrofit 삭제
- Home에서도 기능마다 따로 Retrofit을 뒀길래, Chatting에서도
  (ChattingList, ChattingRoom, ChattingStorage) 따로 두는 편이 좋다고
  생각함.
