+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingRoom/Retrofit/ChattingResponse.kt
@@ -1,6 +1,5 @@
 package com.example.geeksasaeng.Chatting.ChattingRoom.Retrofit
 
-import com.example.geeksasaeng.Home.Delivery.DeliveryBannerResult
 import com.google.gson.annotations.SerializedName
 
 data class CreateChattingRequest(
@@ -88,3 +87,15 @@ data class ChattingDeliveryComplicatedResult(
 
 )
 
+//배달파티 수동 매칭마감
+data class MatchingEndResponse(
+    @SerializedName("code") val code: Int,
+    @SerializedName("isSuccess") val isSuccess : Boolean,
+    @SerializedName("message") val message: String,
+    @SerializedName("result") val result: MatchingEndResult?
+)
+
+data class MatchingEndResult(
+    @SerializedName("deliveryPartyId") val deliveryPartyId: Int,
+    @SerializedName("matchingStatus") val matchingStatus: String
+)
