--- app/src/main/java/com/example/geeksasaeng/Chatting/Retrofit/ChattingView.kt
@@ -1,6 +1,0 @@
-package com.example.geeksasaeng.Chatting.Retrofit
-
-interface MatchingEndView {
-    fun onMatchingEndSuccess()
-    fun onMatchingEndFailure(message: String)
-}
